perf!: replace pyright with python-lsp-server and mypy plugin
BREAKING CHANGE: pyright somehow manages to block the editor
even when just scrolling and leads to blocking at startup.

Until this is fixed I will default to a fast but slightly
less feature-rich setup.
Note, that using the pylsp plugins for mypy and black require
additional installation steps commented in plugisn/quarto.lua
on top of the lspconfig block for pylsp.

mypy currently does not work in quarto documents, because it requires
actual files to work with, while the hidden buffers used for completion
etc. in quarto by otter.nvim are just buffers and currently not written to disk
for performance reasons.